uncomment namechange
The note was added after bmc10011 changed his name but before he used
the new name in a contest.  He used the new name in a contest for the
first time in SotW 576.  As a result of this change, the stats for
bmc10011 and TheRock will be merged into one profile.