adjust by proportion rather than votes
This is one way to ensure that one contest will not entirely determine the ratings of its entrants if a sufficient number of votes is cast. Biggest changes are contests with few votes move ratings *much* more than before. Contests with large numbers of entrants have less aggressive adjustments.

The code works, but the results don't sit well with me... Especially the conservative adjustments in big contests relative to the aggressive adjustments in small contests.